Added WorldTimeSystem
A few things started to become clear here.

1) I didn't like having to update component usages so I am now just
   sending the whole Sim component in SSR and in websockets.

This does reduce the number of websocket messages, but since every component is
already being updated on every tick it's not more HTML.

2) State is not being modeled in an effective way.

It's annoying needing to patch State after it comes across the Worker.
It's also not great needing to update a central State class every time
I need to add more state modules for Systems to work with.

Modeling state might get messier before it gets better.

But hey, at least we have a world date and time now!