NO COMPROMISES!
Got rid of the Worker as it's a premature optimization,
but it also adds extra steps in typing data that crosses.

Modified the State system to be an array of States instead.
I feel like this lets me accumulate state based on the order
of the systems. They can just return the state they own,
and pass the collection to the next one down the line.
This way it's pretty easy to use a discriminated union
to determine what kind of state it is and that even
GIVES TYPE INFORMATION!!

Systems can maintain their own state internally but they
basically have a State DTO to expose to other systems
on the "global" state parameter.

Let's see how this goes.