Simplify Engine for SPEED
Eliminated the State constructor param, since this is configurable
through the System constructors and addSystem methods.

Also removed the Update Callback. This is made possible by
placing a new (async) interval in the index file, and accessing
the state via a new method.

This is measurably faster because the tick doesn't have to wait
for websocket messages to be sent to continue.